\docType{package}
\name{ctsmr}
\alias{ctsmr}
\alias{ctsmr-package}
\alias{package-ctsmr}
\title{Continuous Time Stochastic Modelling for R}
\description{
Continuous Time Stochastic Modelling for R
}
\author{
Rune Juhl \email{ruju@dtu.dk}, Henrik Madsen
\email{hmad@dtu.dk}

Maintainer: Rune Juhl \email{ruju@dtu.dk}
}

