\name{PackData}
\alias{PackData}
\title{Packs (validated) user given data for CTSM}
\usage{
PackData(data, inputs, outputs)
}
\arguments{
  \item{data}{list of data.frames.}

  \item{inputs}{vector of names of the input variables.}

  \item{outputs}{vector of names of the output variables.}
}
\description{
Packs (validated) user given data for CTSM
}

