\name{smooth.ctsmr}
\alias{smooth.ctsmr}
\title{Smoothing filter method for CTSM fits}
\usage{
smooth.ctsmr(object, newdata = NULL, firstorderinputinterpolation = FALSE,
  ...)
}
\arguments{
  \item{object}{Object of class 'ctsmr'}

  \item{newdata}{An optional data frame with new data.}

  \item{firstorderinputinterpolation}{If TRUE (default
  FALSE) first order linear interpolation of the inputs
  between samples.}

  \item{...}{Unused.}
}
\description{
Smoothing filter method for CTSM fits
}

