\name{stochastic.simulate}
\alias{stochastic.simulate}
\title{Stochastic simulation for CTSM fits}
\usage{
stochastic.simulate(fit, x0, data, dt)
}
\arguments{
  \item{fit}{Object of class 'ctsmr'}

  \item{x0}{x0}

  \item{data}{data}

  \item{dt}{Simulation dt for the brownian motion}
}
\description{
Stochastic simulation for CTSM fits
}

